Grace - Built Item and Inventory Classes. Built UseItemRunner and the Input and Output Boundaries. Also included a dummie presenter for the UseItemRunner.
