+++ src/main/java/Mariogatchi/UseItemInputBoundary.java
@@ -1,4 +1,4 @@
-package mariogatchi;
+package Mariogatchi;
 
 public interface UseItemInputBoundary {
