+++ src/main/java/Mariogatchi/UseItemPresenter.java
@@ -1,4 +1,4 @@
-package mariogatchi;
+package Mariogatchi;
 
 public interface UseItemPresenter {
