+++ src/main/java/Mariogatchi/UseItemRequestModel.java
@@ -1,4 +1,4 @@
-package mariogatchi;
+package Mariogatchi;
 
 public class UseItemRequestModel {
     private Item itemToUse;
