+++ src/main/java/Mariogatchi/UseItemResponseModel.java
@@ -1,4 +1,4 @@
-package mariogatchi;
+package Mariogatchi;
 
 public class UseItemResponseModel {
